create delete language endpoint + change logic to update vocabulary item + fix counting average amount

+++ src/models/response/ExerciseResultResponse.php
@@ -7,5 +7,5 @@ class ExerciseResultResponse
     public VocabularyLanguageResponse $language;
     public float $successRate;
     public int $totalAnsweredWords;
-    public int $answeredWordsAverage;
+    public float $answeredWordsAverage;
 }
